Move (and combine) shard index tests
